recsys_and_llm/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0   
scikit-learn==1.6.0
tqdm==4.67.1
pytz==2024.2
accelerate==1.2.1
bitsandbytes==0.45.0
transformers==4.47.1
sentence-transformers==3.3.1
recbole==1.2.0
huggingface_hub==0.27.1
kmeans-pytorch==0.3